api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["farooq-azam-khan <farooq1.khan@ryerson.ca>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
requests = "^2.26.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
email-validator = "^1.1.3"
argon2-cffi = "^21.3.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
